+++ ZezeJava/ZezeJava/src/main/java/Zeze/Netty/HttpBeginStreamHandle.java
@@ -2,5 +2,5 @@ package Zeze.Netty;
 
 @FunctionalInterface
 public interface HttpBeginStreamHandle {
-	void onBeginStream(HttpExchange x, int from, int to, int size) throws Exception;
+	void onBeginStream(HttpExchange x, long from, long to, long size) throws Exception;
 }
